chore: refactor use get staking strategies
